+++ frontend/src/hooks/demo.form-context.ts
@@ -1,0 +1,4 @@
+import { createFormHookContexts } from "@tanstack/react-form";
+
+export const { fieldContext, useFieldContext, formContext, useFormContext } =
+    createFormHookContexts();
