+++ frontend/src/integrations/tanstack-query/layout.tsx
@@ -1,0 +1,5 @@
+import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
+
+export default function LayoutAddition() {
+    return <ReactQueryDevtools buttonPosition="bottom-right" />;
+}
