+++ frontend/src/integrations/tanstack-query/root-provider.tsx
@@ -1,0 +1,17 @@
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+
+const queryClient = new QueryClient();
+
+export function getContext() {
+    return {
+        queryClient,
+    };
+}
+
+export function Provider({ children }: { children: React.ReactNode }) {
+    return (
+        <QueryClientProvider client={queryClient}>
+            {children}
+        </QueryClientProvider>
+    );
+}
