+++ frontend/src/main.tsx
@@ -1,0 +1,48 @@
+import { RouterProvider, createRouter } from "@tanstack/react-router";
+import { StrictMode } from "react";
+import ReactDOM from "react-dom/client";
+
+import * as TanStackQueryProvider from "./integrations/tanstack-query/root-provider.tsx";
+
+// Import the generated route tree
+import { routeTree } from "./routeTree.gen";
+
+import "./styles.css";
+import reportWebVitals from "./reportWebVitals.ts";
+
+// Create a new router instance
+const router = createRouter({
+    routeTree,
+    context: {
+        ...TanStackQueryProvider.getContext(),
+    },
+    defaultPreload: "intent",
+    scrollRestoration: true,
+    defaultStructuralSharing: true,
+    defaultPreloadStaleTime: 0,
+});
+
+// Register the router instance for type safety
+declare module "@tanstack/react-router" {
+    interface Register {
+        router: typeof router;
+    }
+}
+
+// Render the app
+const rootElement = document.getElementById("app");
+if (rootElement && !rootElement.innerHTML) {
+    const root = ReactDOM.createRoot(rootElement);
+    root.render(
+        <StrictMode>
+            <TanStackQueryProvider.Provider>
+                <RouterProvider router={router} />
+            </TanStackQueryProvider.Provider>
+        </StrictMode>,
+    );
+}
+
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+reportWebVitals();
