+++ frontend/src/routes/__root.tsx
@@ -1,0 +1,25 @@
+import { Outlet, createRootRouteWithContext } from "@tanstack/react-router";
+import { TanStackRouterDevtools } from "@tanstack/react-router-devtools";
+
+import Header from "../components/Header";
+
+import TanStackQueryLayout from "../integrations/tanstack-query/layout.tsx";
+
+import type { QueryClient } from "@tanstack/react-query";
+
+interface MyRouterContext {
+    queryClient: QueryClient;
+}
+
+export const Route = createRootRouteWithContext<MyRouterContext>()({
+    component: () => (
+        <>
+            <Header />
+
+            <Outlet />
+            <TanStackRouterDevtools />
+
+            <TanStackQueryLayout />
+        </>
+    ),
+});
