+++ package.json
@@ -1,47 +1,47 @@
-{
-  "name": ".",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "dev": "vite --port 3000",
-    "start": "vite --port 3000",
-    "build": "vite build && tsc",
-    "serve": "vite preview",
-    "test": "vitest run",
-    "format": "biome format",
-    "lint": "biome lint",
-    "check": "biome check"
-  },
-  "dependencies": {
-    "@t3-oss/env-core": "^0.12.0",
-    "@tailwindcss/vite": "^4.0.6",
-    "@tanstack/react-form": "^1.0.0",
-    "@tanstack/react-query": "^5.66.5",
-    "@tanstack/react-query-devtools": "^5.66.5",
-    "@tanstack/react-router": "^1.114.3",
-    "@tanstack/react-router-devtools": "^1.114.3",
-    "@tanstack/router-plugin": "^1.114.3",
-    "class-variance-authority": "^0.7.1",
-    "clsx": "^2.1.1",
-    "lucide-react": "^0.476.0",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
-    "tailwind-merge": "^3.0.2",
-    "tailwindcss": "^4.0.6",
-    "tailwindcss-animate": "^1.0.7",
-    "zod": "^3.24.2"
-  },
-  "devDependencies": {
-    "@biomejs/biome": "1.9.4",
-    "@testing-library/dom": "^10.4.0",
-    "@testing-library/react": "^16.2.0",
-    "@types/react": "^19.0.8",
-    "@types/react-dom": "^19.0.3",
-    "@vitejs/plugin-react": "^4.3.4",
-    "jsdom": "^26.0.0",
-    "typescript": "^5.7.2",
-    "vite": "^6.1.0",
-    "vitest": "^3.0.5",
-    "web-vitals": "^4.2.4"
-  }
-}
+    {
+    "name": "pricepulse",
+    "private": true,
+    "type": "module",
+    "scripts": {
+        "dev": "vite frontend --port 3000",
+        "start": "vite frontend --port 3000",
+        "build": "vite build frontend && tsc",
+        "serve": "vite preview frontend",
+        "test": "vitest run",
+        "format": "biome format --write",
+        "lint": "biome lint --write",
+        "check": "biome check"
+    },
+    "dependencies": {
+        "@t3-oss/env-core": "^0.12.0",
+        "@tailwindcss/vite": "^4.0.6",
+        "@tanstack/react-form": "^1.0.0",
+        "@tanstack/react-query": "^5.66.5",
+        "@tanstack/react-query-devtools": "^5.66.5",
+        "@tanstack/react-router": "^1.114.3",
+        "@tanstack/react-router-devtools": "^1.114.3",
+        "@tanstack/router-plugin": "^1.114.3",
+        "class-variance-authority": "^0.7.1",
+        "clsx": "^2.1.1",
+        "lucide-react": "^0.476.0",
+        "react": "^19.0.0",
+        "react-dom": "^19.0.0",
+        "tailwind-merge": "^3.0.2",
+        "tailwindcss": "^4.0.6",
+        "tailwindcss-animate": "^1.0.7",
+        "zod": "^3.24.2"
+    },
+    "devDependencies": {
+        "@biomejs/biome": "1.9.4",
+        "@testing-library/dom": "^10.4.0",
+        "@testing-library/react": "^16.2.0",
+        "@types/react": "^19.0.8",
+        "@types/react-dom": "^19.0.3",
+        "@vitejs/plugin-react": "^4.3.4",
+        "jsdom": "^26.0.0",
+        "typescript": "^5.7.2",
+        "vite": "^6.1.0",
+        "vitest": "^3.0.5",
+        "web-vitals": "^4.2.4"
+    }
+    }
