--- src/hooks/demo.form-context.ts
@@ -1,4 +1,0 @@
-import { createFormHookContexts } from '@tanstack/react-form'
-
-export const { fieldContext, useFieldContext, formContext, useFormContext } =
-  createFormHookContexts()
