--- src/integrations/tanstack-query/layout.tsx
@@ -1,5 +1,0 @@
-import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
-
-export default function LayoutAddition() {
-  return <ReactQueryDevtools buttonPosition="bottom-right" />
-}
