+++ vite.config.js
@@ -7,14 +7,18 @@ import { resolve } from "node:path";
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [TanStackRouterVite({ autoCodeSplitting: true }), viteReact(), tailwindcss()],
-  test: {
-    globals: true,
-    environment: "jsdom",
-  },
-  resolve: {
-    alias: {
-      '@': resolve(__dirname, './src'),
+    plugins: [
+        TanStackRouterVite({ autoCodeSplitting: true }),
+        viteReact(),
+        tailwindcss(),
+    ],
+    test: {
+        globals: true,
+        environment: "jsdom",
     },
-  }
+    resolve: {
+        alias: {
+            "@": resolve(__dirname, "./src"),
+        },
+    },
 });
